e2e split into parallel and serial flows
