+++ e2e/parallel/7_manualBackup.spec.ts
@@ -6,12 +6,11 @@ import {
   tap,
   tapByText,
   delayTime,
-  tapAtPoint,
   checkIfExistsByText,
   checkIfExists,
   waitAndTap,
   checkIfDoesntExist,
-} from './helpers';
+} from '../helpers';
 
 describe('Backups', () => {
   beforeAll(async () => {
