+++ scripts/run-parallel-e2e.sh
@@ -5,7 +5,7 @@ count=0
 
 until (( count >= max_retries ))
 do
-  ./node_modules/.bin/detox test -c ios.sim.release --maxWorkers 2 -- --forceExit --bail 1
+  ./node_modules/.bin/detox test parallel/ -c ios.sim.release --maxWorkers 4 -- --forceExit --bail 1
   ret_val=$?
   if [ $ret_val -eq 0 ]; then
     exit 0
@@ -14,5 +14,4 @@ do
   echo "Test failed, attempt $count/$max_retries..."
 done
 
-echo "Tests failed after $max_retries attempts."
-exit 1
+
